*: make NowFns heap allocated
Previously NowFns had to be function pointers. This was problematic in
tests which wanted to mock out the NowFn: function pointers can only
refer to globals, but tests run multi-threaded by default and compete
for access to these globals.

Switch NowFn to be a heap-allocated callable. This makes it possible to
supply a closure as the NowFn which refers to local state.

In particular, this means coordtest no longer relies on manipulating
global mutable state. I think in practice coordtest always ran tests
serially, but I think this was incidental, not intentional, and liable
to change inadvertently.