persist: split out a read-only subset of the Blob trait
The first of several PRs to factor out a read-only version of the
persist runtime, for use in the ingestd/dataflowd platform split.

This splits the read-only operations (get/list_keys) of Blob into a new
BlobRead trait (which Blob extends). `close` is primarily responsible
for releasing the exclusive-writer lock and so stays only on Blob.

A `Read` type is added to the Blob trait to establish a relationship
between between an implementation of Blob and its read-only companion.
This relationship isn't used in code, but seems worthwhile to have
around.

A `Config` type is added to the Blob trait and unified
`open_exclusive`/`open_read` constructors are added. These are similarly
not used in code, but seem worthwhile.

An initial version of this PR also had an `as_read` method on Blob that
returned `self` as the corresponding `BlobRead` impl, but given that
Blob extends BlobRead, it was rejected as unnecessary (and it ended up
accidentally being too opinionated about the internals of a Blob impl).