ore: introduce a test::timeout helper
Introduce a test::timeout function that runs a synchronous callback with
a timeout, producing an error if the callback fails to complete within
the specified time.

Then, use this function in test_no_block. The previous approach used a
"time bomb" thread, which panicked if the test hadn't completed within
30s. The defect with this time bomb approach is that it required calling
`ore::panic::set_abort_on_panic` in order for the panic in the time bomb
thread to cause the process to crash. The Rust test runner, however,
relies on catching panics to properly handle error reporting, which
`set_abort_on_panic` interferes with.

The working theory is that `test_no_block` is timing out, but the
`set_abort_on_panic` call is interfering with the test runner's ability
to produce the correct output.

The new `test::timeout` function in ore works around the problem by
runnign the code under test on a separate thread, leaving the "time
bomb" on the main thread, where its panics will be properly noticed by
the test harness.