coord,dataflow: introduce a dataflow::Client trait
Introduce a new trait to abstract over the ways to communicate with a
running dataflow server. The specific implementation of interacting with
a local dataflow server is renamed to dataflow::LocalClient.

This new trait is immediately useful because it allows coordtest to swap
in its own InterceptingDataflowClient implementation without the need
for gross configuration options in the dataflow layer.