persist: split AppliedState out of Indexed
This plumbs BlobCache and Metrics to a bunch of places very temporarily.
Don't worry about it, they'll go back in the next commit.

Also pay no mind to the alternating Indexed and AppliedState impl
blocks. Those will stick around for a bit while we refactor things to
make the changes more obvious, but they'll eventually be cleaned up.