persist: don't consolidate unsealed batches
Previously, we enforced the invariants that data stored in unsealed batches was
- sorted by (time, (key, val))
- consolidated, meaning that each instance of ((key, val), time) occurred exactly
  once.
- diffs were non-zero.

These invariants are potentially valuable in making trace batch construction
more efficient (since records are sorted by time you don't need to scan the entire
unsealed batch), improving the storage space utilization, and holding the data in
a canonical format that makes it trivial to compare if two unsealed batches are
identical. However, we were not using any of these properties currently, and they
were costing us a lot of CPU time in `drain_pending` to enforce, which was directly
contributing to write latency. Also, these properties are harder to enforce in a
columnar representation where its harder to sort and remove records.

This commit removes the three invariants mentioned above, and removes the parts
of drain_pending that enforced them. Unsealed previously used the sortedness
property to derive the min/max time bounds in unsealed batches in constant time,
this is now done with a scan through all the updates in Unsealed::apend. Removing
that additional scan is left for followup work.

indexed_write_drain/mem_unsorted/23000000
                        time:   [231.21 ms 234.27 ms 237.73 ms]
                        thrpt:  [92.268 MiB/s 93.631 MiB/s 94.870 MiB/s]
                 change:
                        time:   [-13.571% -11.868% -10.117%] (p = 0.00 < 0.05)
                        thrpt:  [+11.255% +13.467% +15.702%]
                        Performance has improved.
indexed_write_drain/file_sorted/23000000
                        time:   [323.55 ms 326.15 ms 328.77 ms]
                        thrpt:  [66.716 MiB/s 67.253 MiB/s 67.792 MiB/s]
                 change:
                        time:   [-14.216% -13.266% -12.383%] (p = 0.00 < 0.05)
                        thrpt:  [+14.134% +15.295% +16.571%]
                        Performance has improved.
indexed_write_drain/file_unsorted/23000000
                        time:   [350.38 ms 360.72 ms 369.85 ms]
                        thrpt:  [59.307 MiB/s 60.808 MiB/s 62.602 MiB/s]
                 change:
                        time:   [-15.938% -13.061% -10.478%] (p = 0.00 < 0.05)
                        thrpt:  [+11.704% +15.023% +18.960%]
                        Performance has improved.