coord: only consolidate diffs when some are negative.
We need to consolidate rows in `sequence_send_diffs` to correctly determine
the number of rows affected by a `INSERT`, `UPDATE` or `DELETE` statement.
However, consolidation is only required when there is a mix of positive
and negative diffs, such that some might cancel each other out. When the
diffs are all positive, as is commonly the case in `INSERT`s, the sum of
the absolute values of the diffs is the same regardless of whether the rows are
preconsolidated or not.