deps: upgrade to protobuf v3.0.0-alpha.2
While this is technically an alpha release, the v3 branch of
rust-protobuf is quite robust and contains a few bugfixes not present in
v2. It also contains support for dynamically deserializing messages from
JSON, which will vastly improve the readability of the testdrive tests.

The only downside is that v3 may still have breaking changes, but we'll
have to contend with those eventually anyway.