dataflow: encode source delimited-ness in the type system
This patch adds an associated type to the SourceReader trait that allows
source implementations to choose what type of messages they wish to
produce.

In practice, there are only two valid types depending on whether or not
the source implementation is a delimited source providing a stream of
messages or if it is a non-delimited source which behaves like a stream
of bytes.

The message type for delimited source is an `Option<Vec<u8>>` whereas
for the bytestream ones the pre-existing `MessagePayload` enum is used.