persist: make thread names more idiomatic
These better match tokio and timely, both of which use ":" to delineate
an initial category and a "work-[num]" suffix as the worker thread
names. Also remove the persist runtime id from the thread name, which
wasn't getting us anything any might get in the way when graphing (since
it would change on each restart).