persist: change Blob::set allow_overwrite arg to be require_atomic
There doesn't seem to be any way to performantly implement
allow_overwrite=false for s3. Conveniently, after the keys for batch
blobs were changed to UUIDs in #8667, we no longer need the stronger
semantics of allow_overwrite=false to guard against accidental
overwrites. We do however need the ability to opt-in to atomic writes in
FileBlob for the meta key, so keep this around in a weakened form as
require_atomic.