persist: fix source bindings restore
Before, we would restore bindings that are beyond the common seal
timestamp but not keep the information that we know of the partition.
This would lead to panics when trying to add the binding to
`timestamp_histories` at the call site of restore().

Now, we record that we know of a partition without restoring the offset,
by returning a Some/None for the restored offsets from restore().

Fixes #8909