calculate schemas of decode and envelope stages separately
Previously, when we wanted to compute the `RelationDesc` that a
particular source would produce we relied solely on the
`SourceDataEncoding::desc` method. This method was in theory only
concerned about the decoding stage, but in practice it was being passed
as a parameter the resulting envelope which was being used to compute
the correct relation type.

This patch makes the computation of the RelationDesc of a source in two
distinct steps.

First, we perform the calculation on the decoding stage which produces a
`(Option<RelationDesc>, RelationDesc)` pair which has a direct
corresondence to the `DecodeResult` output of the decoding stage.

Then, the envelopes are asked about their `RelationDesc` output, given a
particular decoding stage input. This is where each envelope
implementation is responsible to transform their input type into an
output type. The most simple example is the `NONE` envelope with no
`INCLUDE KEY` directive which simply declares its output to be identical
to the value relation description.

This brings us closer to make the decoding and envelope processing
stages work purely on Rows and not allow any intermingling of decoding
formats and envelopes. Concretely this means that we should be able to
use any envelope with any format so long as the input and output
`RelationDesc`s match.

Finally, a new feature fell out of this refactor which is that we now
support ingesting avro schemas that their top node is not a record. We
were previously rejecting such schemas for the general case but
admitting them for CDCv2 envelopes which have their top node be a union.

In order to close the gap and fix this inconsistency we now create a
relation with a single non-nullable column whose type is the type of the
top node when we encounter a schema that's not a record.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>