expr: simplify the function definition macro
This is an attempt to make the function generation macro a lot simplier
by relying a lot more on traits and the type system.

The primary motivation for this change is to make implementing eager
functions *without* the macro also be reasonably short and
straightforward.

Implementing SQL function with the EagerUnaryFunc directly is needed for
those UnaryFunc variants that hold data, and therefore cannot work with
the generated ZST type.

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>