persist/golden_test: print out current state if we cannot parse golden state
Some changes to the encoding state render the previously stored golden state
entirely un-decodable. In those cases, the golden test currently crashes with
an assertion failure (it returns an error instead of the expected ok). This commit
makes the behavior in that case a little bit more user friendly by printing out
the current state and the requests that lead to it, as most of the time the in
this case we just want to update golden_test.json.