persist: introduce a ColumnarRecords object that store updates in a columnar format.
At a very high level, our data is composed of lots of small vectors of bytes for
keys and values. This representation turns the storage around 90 degrees to be
backed by one large vector of bytes for all the keys + another large vector of
bytes for all the values, and separate arrays of offsets into both.

The pros of this representation are that its faster to encode, and more performant
to scan through (fewer cache misses). The cons are that its more complicated to
append data, and its downright difficult to reorder or shuffle data in this
representation.

Actually using this representation is left to future commits.