ore: tolerate larger stacks in debug builds
@petrosagg tracked down why several functions have such massive stacks
in debug builds. Details within the patch, but the tl;dr is that the
large match statements can result in huge stack frames when compiler
optimizations are disabled.

So, this commit increases the `STACK_RED_ZONE` and `STACK_SIZE` values
in debug builds. That way `stack::maybe_grow` actually has a chance to
notice when the stack is running low on space in debug builds.

Note that we don't intend to worry much about excessively nested SQL
queries causing stack overflows in debug builds, since e.g. the lack of
tail call optimization can cause auto generated `Clone` implementations
to overflow. But this particular issue was causing very reasonable SQL
statements (see #9121) to overflow the stack in debug builds.