sql: dispatch to a separate function when planning Expr variants
When building materialize in debug mode the compiler is not smart enough
to re-use stack space between all mutually exclusive paths of execution
that lead to each individual enum variant. This has the effect of
calculating a stack size for the function that is the *sum* of all local
variables.

The function in question was calculated to have a 32kB stack size in
debug builds which which adds up quickly when recursing.

The changes of this patch bring this number down to ~7kB.

Related to #9121

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>