doc/user: update release versioning scheme
Our current release versioning scheme is empirically confusing. External
users often think that a release like v0.8.1 only contains minor bug
fixes on top of v0.8.0, when in fact it can contain major new features.
We've made this mistake internally, too: we've written blog posts that
talk about new features in v0.9.0, when in fact many of the described
features were released in v0.8.1, v0.8.2, and so on.

And, in fact, it would be nice to have a scheme that *could* indicate
when a release only contains bug fixes. When we need to cut an emergency
release for a critical bug or security vulnerability, we should use the
version number to indicate that the release contains only the bug fix
and no other breaking changes to incentivize users to upgrade more
quickly than they would for a normal release.

So, the proposal here is to switch to a more standard versioning scheme.
Weekly releases would increment the middle component (v0.10.0, v0.11.0,
v0.12.0, etc.) to make space for emergency bugfix releases that bump the
last component (e.g. v0.10.1).

The downside is that we lose the clear moment at which to write a
release blog post announcing the new features. Previously, every
milestone release (e.g., v0.8.0, v0.9.0) was accompanied by a blog post.
But most weekly releases are relatively boring and not worthy of a blog
post.

I propose that instead we couple blog posts to the forthcoming monthly
milestones. Each month, we can write a "What's new in Materialize" blog
post that describes the highlights that shipped in that month's
releases. This approach pairs better with Materialize Cloud, too, where
releases happen multiple times per day.