persist: in conditional_seal(), don't allow frontier progress on failing seals
We hold on to capabilities by storing them along with the seal futures
in `pending_futures`. This blocks the downstream frontier from advancing
while we are still waiting for a seal to be successful.

Before, we were letting go of the capability both when the future
finishes sucessfully and when it errored. This is incorrect because a
failed seal should not allow downstream progress.

Now, we retry a seal when it fails, with the same timestamp.
Additionally, we remove pending seal futures when they are superseded by
a successful seal for a later timestamp. The motivation behind this
latter part is that it can happen that a seal will never succeed,
because it is "before" the timestamp up to which the collection has
already been sealed. Eventually, some later seal will succeed and we can
then let go of this failing seal.

Additionally, we downgrade log::error! to log::trace! when logging seal
errors because we are handling them internally and an ERROR message in
the log would potentially concern users.

Fixes #9241
Fixes #9189 because we downgrade seal errors to trace! logging and don't
emit errors anymore.