persist: in conditional_seal(), don't emit errors when sealing fails
Before, we would emit errors to the "err" stream, which could wedege a
view if they were wired into the views error collection.

We previously changed the the seal operator to 1) retry seals on
failure, and 2) drop seal futures when superseded by a later successful
seal. This turns seal errors into transient errors that should not wedge
a view.