sql: allow multiple table functions as long as they are identical
When the same table function appears multiple times in a SELECT
projection, we can replace all instances with an identifier to an
additional FROM clause. This allows us to visit the expression tree and
support more nested expressions, even ones that want only a specific
column from a multi-column table function.

The unsupported error referencing #1546 can be moved to the single place
where we still don't fully support this (multiple table functions in
SELECT projections. It is converted to a regular error in all other
places, like in a WHERE clause, matching Postgres.

Informs #1546