sql: support OPERATOR syntax and default `pg_catalog` namespace (#9255)
Add support for `OPERATOR(<namespace>.<operator>)`. 

Currently, Materialize does not support the creation of custom operators and this PR does not attempt to add it. The `<namespace>` is optional and can be omitted. The only supported namespace is `pg_catalog`, all other namespaces will be rejected.

The `<operator>` can be any operator such as `+`, `*` or `~`.

The precedence of OPERATOR(<op>) is different from the bare operator and follows the PG precedence: https://www.postgresql.org/docs/14/sql-syntax-lexical.html#SQL-PRECEDENCE