persist: use protobuf for metadata encoding
We initially used Abomonation and then serde+bincode for convenience and
ease of development, but neither was ever intended as the long-term
solution for how to encode metadata.

This switches to Google's Protocol Buffers version 3 (aka [protobuf],
proto), which has a considerable number of hours of dealing with
backward and forward compatibility under its belt. Other options that
were seriously considered were Cap'n Proto and Flatbuffers.

[protobuf]: https://developers.google.com/protocol-buffers

For ease of review, this commit introduces a very thin shim that uses a
protobuf equivalent of our previously encoded structs as a middleman for
encoding and decoding. As a result, most persist code doesn't have to
change. Still TODO is figuring out whether we would like to keep this
pattern or move more toward serializing and deserializing the persist
state directly to/from the generated protobuf structs. Each has various
advantage and disadvantage tradeoffs that have yet to be fully explored.

I'm skeptical of these benchmark improvements, but they seem repeatable.

    indexed_write_drain/mem_sorted/23000000
                            time:   [142.54 ms 145.81 ms 147.92 ms]
                            thrpt:  [148.29 MiB/s 150.43 MiB/s 153.88 MiB/s]
                     change:
                            time:   [-7.8843% -5.9817% -3.9180%] (p = 0.00 < 0.05)
                            thrpt:  [+4.0778% +6.3622% +8.5591%]
                            Performance has improved.
    indexed_write_drain/mem_unsorted/23000000
                            time:   [147.02 ms 147.99 ms 149.98 ms]
                            thrpt:  [146.25 MiB/s 148.22 MiB/s 149.19 MiB/s]
                     change:
                            time:   [-6.5835% -5.1494% -3.5387%] (p = 0.00 < 0.05)
                            thrpt:  [+3.6686% +5.4290% +7.0475%]
                            Performance has improved.
    indexed_write_drain/file_sorted/23000000
                            time:   [208.09 ms 210.09 ms 212.47 ms]
                            thrpt:  [103.24 MiB/s 104.40 MiB/s 105.41 MiB/s]
                     change:
                            time:   [-2.2260% -0.0583% +2.2228%] (p = 0.96 > 0.05)
                            thrpt:  [-2.1745% +0.0583% +2.2767%]
                            No change in performance detected.
    indexed_write_drain/file_unsorted/23000000
                            time:   [209.71 ms 211.21 ms 212.55 ms]
                            thrpt:  [103.20 MiB/s 103.85 MiB/s 104.59 MiB/s]
                     change:
                            time:   [+0.1669% +2.1178% +3.6465%] (p = 0.04 < 0.05)
                            thrpt:  [-3.5182% -2.0739% -0.1666%]
                            Change within noise threshold.