persist: change write microbenchmarks to generate ColumnarRecords directly
Previously, the various write benchmarks would generate a large vector of
((key, val), time, diff) records which would then be converted into a `ColumnarRecords`
object if needed. This forced the benchmarks which didn't need the underlying
vector of records to still pay the allocation cost on benchmark setup.

This commit changes the behavior to instead always generate a `ColumnarRecords`
and downgrade to a vector of records only as needed. This change also removes
the clone needed to find the encoded length of a set of records.

This commit has no significant impact on benchmark results.