persist: fix error on empty writes
After we switched to a columnar encoding, persist started erroring on empty write
commands. Previously, we would correctly detect when the set of pending writes was
empty and simply refuse to write a new unsealed batch. The commit that added columnar
unsealed batches broke the emptiness check, because we replaced it with a check
that the `Vec<ColumnarRecords>` itself was empty, which is more like checking that
no writes were performed.

This commit changes the behavior to only store non-empty ColumnarRecords in `pending`,
and adds a sanity check for that invariant plus a regression test. We chose to drop
the empty records as early as possible, because we don't want to be encoding these
empty ColumnarRecords at all its a pure waste of storage space, so they would either
have to be dropped just before we encode an UnsealedBatch at the latest, and from
that, there didn't seem to be any reason to keep them around at all.