coord, dataflow: Move TAIL row processing to coord
Currently, the dataflow layer prepares the response to a TAIL command, in
which it creates rows encoding data and optional progress updates. This
is problematic as the coordinator needs to extract metadata from rows
instead of structured information if it wishes to do so. With this change,
the formulation of rows sent to the client is moved to the coordinator.

Signed-off-by: Moritz Hoffmann <mh@materialize.com>