persist: make create_or_load() infallible
We now keep the error that create_or_load() would return internally in
`StreamWriteHandle`/`StreamReadHandle` and raise it when their methods
are called. This delays surfacing the error to the first time one of the
methods are called.

We need this in order to make dataflow graph construction deterministic
for Tables and persisted Sources. Currently, the graph shape differs
depending on wether this call succeeds or not, so it can happen that one
worker creates a graph shape that is different from the others, which...
big no-no for timely dataflow. With this change, we are not completely
there yet because we also need to move getting the upper seal timestamp
out of graph construction. It is a necessary first step, though.

The thing this changes on the interfaces is that `create_or_load()` is
now infallible while `stream_id()` becomes fallible because we only have
a stream id when initialization of the handle succeeds. Method
`stream_name()` is still infallible, though I could change that to be
fallible as well.