persist: refuse to read data from the future
At the moment, we can still only detect and error on a forward
compatibility violation by bumping the ENCODING_VERSION, which we'll
have to stop doing when we start guaranteeing backward compatibility.
This commit saves mz build info in ProtoMeta. Since this is the
entrypoint for all written persist metadata and data, it's an upper
bound on versions involved in any persist data. Then, whenever a
ProtoMeta is read in (currently only at startup and in debug_assertion
guarded validations), it checks the written version vs the running code.