sql: structure more errors
Plumb `PlanError` through more of the SQL planner. In many cases we bail
out to `PlanError::Unstructured` immediately, but this sets us up to
gradually structure individual calls to `sql_bail!` as described in #5340.

In particular, this sets us up to tackle #5314, where the SQL planner
wants to special-case UnknownColumn errors that occur while planning
SELECT lists to provide a friendlier error message.